wireguard: do not overwrite an existing /etc/gai.conf file

+++ trunk/user/wireguard/wgc.sh
@@ -115,7 +115,7 @@ setconf_wg()
         PEER_ALLOWEDIPS=$(echo "$PEER_ALLOWEDIPS" | tr -s ',' '\n' | grep -v ':' | tr -s '\n' ',' | sed 's/,$//')
     fi
 
-    local awg
+    local awg gai
     if [ "$MODULE" = "amneziawg" ]; then
         cps()
         {
@@ -148,12 +148,13 @@ EOF
     [ "$IF_PRESHARED" ] && echo "PresharedKey = $IF_PRESHARED" >> "/tmp/${IF_NAME}.conf.$$"
 
     # increase the priority of ipv4
-    echo "precedence ::ffff:0:0/96  100" > /etc/gai.conf
+    [ -s /etc/gai.conf ] && gai=1
+    [ -z "$gai" ] && echo "precedence ::ffff:0:0/96  100" > /etc/gai.conf
 
     log_try_connect
     local res=$($WG setconf $IF_NAME "/tmp/${IF_NAME}.conf.$$" 2>&1)
 
-    rm -f /etc/gai.conf
+    [ -z "$gai" ] && rm -f /etc/gai.conf
     rm -f "/tmp/${IF_NAME}.conf.$$"
 
     [ "$1" = "reconnect" ] && return
